requirements.txt
google-api-python-client
oauth2client
langchain
openai
python-dotenv
mendable-py